Debug destroy error

+++ lib/player.py
@@ -2,7 +2,7 @@
 """
 Плеер это чувак, который тилибонькает фигурки.
 Плеер = игровое пространство плеера, в котором бултыхаются фигурки. И бобер.
-Плеер связан со стеной, которая связывает его игровое пространство с 
+Плеер связан со стеной, которая связывает его игровое пространство с
 пространством каким-то другим."""
 import uuid
 from .figure import new_figure, delete_figure
@@ -27,19 +27,23 @@ def new_player(name='', figures=[], uid=None):
 
 
 def destroy_player(uid):
-    player = get_player(uid)
-    if not player:
-        return
-    for fig in player['figures']:
-        delete_figure(fig)
-    for wall_key in redis_walls.keys():
-        wall = get_redis_value(wall_key, redis_walls)
-        print('TRYING TO DELETE PLAYERS', wall['players'], uid, uid in wall['players'])
-        if uid in wall['players']:
-            wall['players'].remove(uid)
-            set_redis_value(wall['uid'], wall, redis_walls)
-    redis_players.delete(uid)
-    print('DELETED PLAYER', uid)
+    print("STARTED DESTROING", uid)
+    if (uid):
+        player = get_player(uid)
+        if not player:
+            return
+        for fig in player['figures']:
+            delete_figure(fig)
+        for wall_key in redis_walls.keys():
+            wall = get_redis_value(wall_key, redis_walls)
+            print('TRYING TO DELETE PLAYERS', wall['players'], uid, uid in wall['players'])
+            if uid in wall['players']:
+                wall['players'].remove(uid)
+                set_redis_value(wall['uid'], wall, redis_walls)
+        redis_players.delete(uid)
+        print('DELETED PLAYER', uid)
+    else:
+        print("NO UID PROVIDED FOR DESTROING PLAYER")
 
 
 def remove_figure(player_uid, figure_uid):
@@ -55,5 +59,3 @@ def add_figure(player_uid, figure_uid):
     player = get_redis_value(player_uid, redis_players)
     player['figures'].append(figure_uid)
     set_redis_value(player_uid, player, redis_players)
-
-
